feat: la fiche produit porte sa propre note, plus celle de la boutique
T022 à T033. `ratingAvg` et `ratingCount` traversent le contrat produit,
le résumé, la recherche et le mappage mobile. La fiche affiche une ligne
compacte près du prix et une bande « Avis » — moyenne, répartition en
cinq barres, trois avis récents, puis « Voir les N avis ».

La note de la boutique quitte la fiche produit (FR-009). Deux notes
voisines empêchent de savoir laquelle lire, et la fiche fournisseur
affiche déjà les siennes.

Un bug réel trouvé en faisant tourner le calcul sur de vraies données :
la moyenne pondérée divisait en **entier**. (5×2 + 4×2 + 3×1) / 5 = 21/5
rendait 4 au lieu de 4,2. La version boutique y échappe par hasard, son
`/ 4.0` rendant l'expression flottante. Cast `::numeric` ajouté, et un
test verrouille la présence du cast.

`formatRelativeDate` sort de `reviews-list.tsx`, où il était local, vers
`features/common/format-date.ts` : les avis produits datent comme les
avis de boutique, sans duplication.

219 tests au vert, lint sans erreur.

+++ apps/api/src/modules/ratings/product-reviews.service.ts
@@ -200,14 +200,17 @@ export class ProductReviewsService {
    * Recomputes a product's average and count from its visible reviews.
    *
    * Same weighting as `RatingsService.recalculateRating`, minus the four
-   * criteria: the rating is already a single number. The average stays null
+   * criteria: the rating is already a single number. Hence the `::numeric`
+   * cast — the shop version divides by `4.0` and gets its float for free,
+   * whereas summing smallints over integers here would divide as integers
+   * and turn an average of 4.2 into 4. The average stays null
    * below the threshold, which drops the product to the end of a `NULLS LAST`
    * ordering with no extra code.
    */
   async recalculateProductRating(productId: string): Promise<void> {
     const rows = await this.em.getConnection().execute(
       `SELECT COALESCE(
-        SUM(rating * CASE WHEN "createdAt" >= NOW() - INTERVAL '${RECENT_WINDOW_DAYS} days' THEN 2 ELSE 1 END)
+        SUM(rating::numeric * CASE WHEN "createdAt" >= NOW() - INTERVAL '${RECENT_WINDOW_DAYS} days' THEN 2 ELSE 1 END)
         / NULLIF(SUM(CASE WHEN "createdAt" >= NOW() - INTERVAL '${RECENT_WINDOW_DAYS} days' THEN 2 ELSE 1 END), 0),
         0
       ) as weighted_avg,
